+++ data.base.api/src/test/java/ru/tinkoff/qa/neptune/data/base/test/persistable/object/operations/BaseDbOperationTest.java
@@ -1,7 +1,7 @@
 package ru.tinkoff.qa.neptune.data.base.test.persistable.object.operations;
 
 import org.testng.annotations.BeforeSuite;
-import ru.tinkoff.qa.neptune.data.base.api.DataBaseStepContext;
+import ru.tinkoff.qa.neptune.data.base.api.DataBaseStepContextDeprecated;
 
 import java.io.File;
 import java.io.IOException;
@@ -19,7 +19,7 @@ public abstract class BaseDbOperationTest {
     private static final String TEST_BASE2 = "TestBase2";
     static File testDB1;
     static File testDB2;
-    protected static DataBaseStepContext dataBaseSteps;
+    protected static DataBaseStepContextDeprecated dataBaseSteps;
 
     @BeforeSuite
     public void copyDB() throws IOException {
@@ -30,7 +30,7 @@ public abstract class BaseDbOperationTest {
         Files.copy(db2.toPath(), (testDB2 = new File(format("%s.db", TEST_BASE2))).toPath(), REPLACE_EXISTING);
         System.setProperty(DEFAULT_JDO_PERSISTENCE_MANAGER_FACTORY_PROPERTY.getPropertyName(),
                 ConnectionDataSupplierForTestBase1.class.getName());
-        dataBaseSteps = getProxied(DataBaseStepContext.class);
+        dataBaseSteps = getProxied(DataBaseStepContextDeprecated.class);
         testDB1.deleteOnExit();
         testDB2.deleteOnExit();
     }
